Fixes #33.
We now display "Joystick not plugged in" in grey if we don't see one, and we check for the presence of at least one once per frame.  The moment we see one, we go back to normal menu operation.

Gameplay already does the right thing, since it just calls Global.poll_joystick() which polls keyboard as well, but, as an enhancement, we should either pause (once that's implemented) or display a message telling the user the joystick's gone, reverting to keyboard input.